sweep2sleep: Adjust to new touchscreen driver implementation of OOS3 (V1.1.1)
Issue was, a double tap on status bar to put the device asleep was having a side effect
after next wake up by double tap, it was putting the device asleep straight away again.

This is due to the fact, the driver does not report updated coordinates anymore when
screen is off, which triggered a wrong double tap to sleep again with the old coordinates
of last double tap to sleep.